*** empty log message ***
git-svn-id: http://svn.neuromancy.net/magick/trunk/Magick-II@451 1d68a2a4-f605-0410-a2a0-863aa388aa54

+++ src/datetime.cpp
@@ -828,6 +828,6 @@ int mDateTime::DaysSince()
 
 int mDateTime::YearsSince()
 {
-    return (DaysSince() / 365.5);
+    return (int)((double) DaysSince() / 365.5);
 }
